api/pyproject.toml
[tool.poetry]
name = "greenhouse-monitor-api"
version = "0.1.0"
description = "AI-powered greenhouse monitoring system API"
authors = ["Zeynep Sariteke"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.6"
openai = "^1.54.0"
python-dotenv = "^1.0.0"
pillow = "^10.2.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-json-logger = "^2.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.12.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
